+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.2] 2023-03-10
+
 ### Added
 
 - [Document building farmOS with Composer #648](https://github.com/farmOS/farmOS/pull/648)
@@ -378,7 +380,8 @@ moving forward.
     Drupal 7, which required a complete refactor of the codebase. By comparison,
     updating from Drupal 9 to 10 will simply involve updating deprecated code.
 
-[Unreleased]: https://github.com/farmOS/farmOS/compare/2.0.1...HEAD
+[Unreleased]: https://github.com/farmOS/farmOS/compare/2.0.2...HEAD
+[2.0.2]: https://github.com/farmOS/farmOS/releases/tag/2.0.2
 [2.0.1]: https://github.com/farmOS/farmOS/releases/tag/2.0.1
 [2.0.0]: https://github.com/farmOS/farmOS/releases/tag/2.0.0
 [2.0.0-rc1]: https://github.com/farmOS/farmOS/releases/tag/2.0.0-rc1